automation/requirements.txt
kiteconnect==5.0.1
pyotp==2.9.0
python-dotenv==1.0.1
requests==2.32.3
selenium==4.19.0
webdriver-manager==4.0.1
